+++ Command/PeekTubeCommand.php
@@ -4,12 +4,13 @@ namespace Leezy\PheanstalkBundle\Command;
 
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 use Pheanstalk_Exception;
 
-class StatsCommand extends ContainerAwareCommand
+class PeekTubeCommand extends ContainerAwareCommand
 {
     /**
      * @see Command
@@ -17,16 +18,22 @@ class StatsCommand extends ContainerAwareCommand
     protected function configure()
     {
         $this
-            ->setName('leezy:pheanstalk:stats')
+            ->setName('leezy:pheanstalk:peek-tube')
+            ->addArgument('tube', InputArgument::REQUIRED, 'The tube to peek.')
+            ->addOption('buried', 'b', InputOption::VALUE_NONE, 'Peek in buried instead of ready')
             ->addArgument('pheanstalk', InputArgument::OPTIONAL, 'Pheanstalk name.')
-            ->setDescription('Gives statistical information about the beanstalkd system as a whole.');
+            ->setDescription('Take a peek at the first job in a tube, ready or burried.');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        $tube = $input->getArgument('tube');
+        $buried = $input->getOption('buried');
+
         $pheanstalkName = $input->getArgument('pheanstalk');
 
         $pheanstalkLocator = $this->getContainer()->get('leezy.pheanstalk.pheanstalk_locator');
+        /** @var \Pheanstalk_Pheanstalk $pheanstalk */
         $pheanstalk = $pheanstalkLocator->getPheanstalk($pheanstalkName);
 
         if (null === $pheanstalkName) {
@@ -46,22 +53,21 @@ class StatsCommand extends ContainerAwareCommand
         }
 
         try {
-            $stats = $pheanstalk->stats();
-
-            if (count($stats) === 0 ) {
-                $output->writeln('Pheanstalk : <error>' . $pheanstalkName . '</error>');
-                $output->writeln('<info>0 stats.</info>');
-
-                return;
+            if ($buried) {
+                $job = $pheanstalk->peekBuried($tube);
+            } else {
+                $job = $pheanstalk->peekReady($tube);
             }
 
-            $output->writeln('Pheanstalk : <info>' . $pheanstalkName . '</info>');
-
-            foreach ($stats as $key => $information) {
-                $output->writeln('- <info>' . $key . '</info> : ' . $information);
+            if ($job) {
+                $output->writeln(sprintf('Pheanstalk : <info>%s</info>', $pheanstalkName));
+                $output->writeln(sprintf('Tube : <info>%s</info>', $tube));
+                $output->writeln(sprintf('Job id : <info>%s</info>', $job->getId()));
+                $output->writeln(sprintf('Data : <info>%s</info>', $job->getData()));
             }
         } catch (Pheanstalk_Exception $e) {
-            $output->writeln('Pheanstalk : <info>' . $pheanstalkName . '</info>');
+            $output->writeln(sprintf('Pheanstalk : <info>%s</info>', $pheanstalkName));
+            $output->writeln(sprintf('Tube : <info>%s</info>', $tube));
             $output->writeln(sprintf('<error>%s</error>', $e->getMessage()));
         }
     }