+++ DependencyInjection/LeezyPheanstalkExtension.php
@@ -9,8 +9,6 @@ use Symfony\Component\DependencyInjection\Loader;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 
-use Leezy\PheanstalkBundle\Exceptions\PheanstalkException;
-
 /**
  * This is the class that loads and manages your bundle configuration
  *
@@ -18,18 +16,6 @@ use Leezy\PheanstalkBundle\Exceptions\PheanstalkException;
  */
 class LeezyPheanstalkExtension extends Extension
 {
-    protected function reservedName()
-    {
-        return array(
-                'pheanstalks',
-                'pheanstalk_locator',
-                'proxy',
-                'data_collector',
-                'listener',
-                'event',
-            );
-    }
-
     /**
      * {@inheritDoc}
      */
@@ -64,47 +50,7 @@ class LeezyPheanstalkExtension extends Extension
         // Create a connection locator that will reference all existing connection
         $connectionLocatorDef = new Definition("Leezy\PheanstalkBundle\PheanstalkLocator");
         $container->setDefinition("leezy.pheanstalk.pheanstalk_locator", $connectionLocatorDef);
-
-        $defaultPheanstalkName = null;
-        $pheanstalks = $config['pheanstalks'];
-
-        $pheanstalkLocatorDef = $container->getDefinition("leezy.pheanstalk.pheanstalk_locator");
-
-        // For each connection in the configuration file
-        foreach ($pheanstalks as $name => $pheanstalk) {
-            if (in_array($name, $this->reservedName())) {
-                throw new \RuntimeException('Reserved pheanstalk name : ' . $name);
-            }
-
-            $pheanstalkConfig = array($pheanstalk['server'], $pheanstalk['port'], $pheanstalk['timeout']);
-            $isDefault = $pheanstalk['default'];
-
-            $pheanstalkDef = $container->getDefinition($pheanstalk['proxy']);
-
-            //TODO: Add Reflection to check PheanstalkProxyInterface implementation
-            //$pheanstalkRefl = new \ReflectionClass($pheanstalkDef->getClass());
-            //$pheanstalkRefl->implementsInterface('Leezy\PheanstalkBundle\Proxy\PheanstalkProxyInterface')
-            $pheanstalkDef->addMethodCall('setPheanstalk', array(new Definition('Pheanstalk_Pheanstalk', $pheanstalkConfig)));
-            $pheanstalkDef->addMethodCall('setName', array($name));
-
-            $container->setDefinition("leezy.pheanstalk." . $name, $pheanstalkDef);
-
-            // Register the connection in the connection locator
-            $pheanstalkLocatorDef->addMethodCall('addPheanstalk', array(
-                $name,
-                $container->getDefinition("leezy.pheanstalk." . $name),
-                $isDefault
-            ));
-
-            if ($isDefault) {
-                if (null !== $defaultPheanstalkName) {
-                    throw new PheanstalkException(sprintf("Default pheanstalk already defined. '%s' & '%s'", $defaultPheanstalkName, $name));
-                }
-
-                $defaultPheanstalkName = $name;
-                $container->setAlias("leezy.pheanstalk", "leezy.pheanstalk." . $name);
-            }
-        }
+        $container->setParameter('leezy.pheanstalk.pheanstalks', $config['pheanstalks']);
     }
 
     /**
@@ -122,13 +68,5 @@ class LeezyPheanstalkExtension extends Extension
         $dataCollectorDef->addArgument(new Reference('leezy.pheanstalk.pheanstalk_locator'));
 
         $container->setDefinition("leezy.pheanstalk.data_collector", $dataCollectorDef);
-    }
-
-    public function process(ContainerBuilder $container)
-    {
-        if (!$container->hasParameter('leezy.pheanstalk.pheanstalks')) {
-            return;
-        }
-
     }
 }